test/requirements.txt
pandas==2.2.3
python-dotenv==1.1.0
supabase==2.15.0
